+++ pyproject.toml
@@ -28,8 +28,7 @@ dependencies = [
     "semver>=2.13.0,<4.0",
     "six>=1.16.0,<2.0.0",
     "tqdm>=4.59.0,<5.0.0",
-    "werkzeug>=2.3.7,<4.0.0",
-    "requests>=2.28.2,<3",
+    "werkzeug>=2.3.7,<4.0.0"
 ]
 
 [project.scripts]
@@ -54,6 +53,10 @@ runtime-http = [
     "sse-starlette>=1.6.1,<2",
 ]
 
+runtime-client = [
+    "requests>=2.28.2,<3",
+]
+
 interfaces-vision = [
     "pillow>=6.2.1,<11.0"
 ]
@@ -70,7 +73,7 @@ interfaces-ts-pyspark = [
 
 # NOTE: This is "all" from the user perspective, not the dev perspective
 all = [
-    "caikit[runtime-grpc, runtime-http, interfaces-vision, interfaces-ts]",
+    "caikit[runtime-grpc, runtime-http, runtime-client, interfaces-vision, interfaces-ts]",
 ]
 
 ## Dev Extra Sets ##
@@ -84,7 +87,7 @@ dev-test = [
     "pytest>=6.2.5,<8.0",
     "tls_test_tools>=0.1.1",
     "wheel>=0.38.4",
-    "caikit[interfaces-vision, interfaces-ts-pyspark]",
+    "caikit[interfaces-vision, interfaces-ts-pyspark, runtime-client]",
 ]
 
 dev-docs = [
